requirements.txt
joblib~=1.4.2
transformers~=4.46.3
numpy~=2.1.3
pandas~=2.2.3
scikit-learn~=1.5.2
sentence-transformers~=3.3.1
scipy~=1.14.1
Flask~=3.1.0
sounddevice~=0.5.1
ffmpeg~=1.4
ffmpeg-python~=0.2.0
PyMySQL~=1.1.1
python-dotenv~=1.0.1
requests~=2.32.3
spotipy~=2.24.0
torch~=2.5.1
SQLAlchemy~=2.0.36
gunicorn~=22.0.0
google-generativeai~=0.8.4